refactor: separate AuthToken and Category interfaces
