+++ frontend/src/model/AuthToken.ts
@@ -1,0 +1,3 @@
+export interface AuthToken {
+    token: string;
+}
