+++ frontend/src/model/Category.ts
@@ -1,0 +1,4 @@
+export interface Category {
+  id: number;
+  name: string;
+}
