--- frontend/src/model/model.ts
@@ -1,8 +1,0 @@
-export interface AuthToken {
-    token: string;
-}
-
-export interface Category {
-    id: number;
-    name: string;
-}
