+++ frontend/src/utils/mockAuthToken.ts
@@ -1,4 +1,4 @@
-import { AuthToken } from "@/model/model";
+import { AuthToken } from "@/model/AuthToken";
 
 export const mockAuthToken: AuthToken = {
   token: "FAKE-AUTH-TOKEN"
