+++ frontend/test/service/AuthService.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, it } from "@jest/globals";
 import AuthenticationService from "@/service/AuthenticationService";
 import IServerFacade from "@/service/IServerFacade";
 import { instance, mock, verify } from "@typestrong/ts-mockito";
-import { AuthToken } from "@/model/model";
+import { AuthToken } from "@/model/AuthToken";
 
 const EMAIL = "test.email@email.com";
 const PASS = "test-password";
