+++ frontend/test/service/ContactService.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, it } from "@jest/globals";
 import IServerFacade from "@/service/IServerFacade";
 import { instance, mock, verify } from "@typestrong/ts-mockito";
-import { AuthToken } from "@/model/model";
+import { AuthToken } from "@/model/AuthToken";
 import ContactService, { NewContactData } from "@/service/ContactService";
 import { Contact } from "@/utils/mockContacts";
