+++ frontend/test/service/TimelineService.test.tsx
@@ -1,7 +1,7 @@
 import { beforeEach, describe, it } from "@jest/globals";
 import IServerFacade from "@/service/IServerFacade";
 import { anything, instance, mock, verify, when } from "@typestrong/ts-mockito";
-import { AuthToken } from "@/model/model";
+import { AuthToken } from "@/model/AuthToken";
 import EventService, { TimelineQuery } from "@/service/EventService";
 
 const QUERY_W_NO_REQUIREMENTS: TimelineQuery = {
